Moved tests and fixed game list subheader
