+++ app/tests/boilerplate/i18n.test.js
@@ -1,6 +1,6 @@
-import { formatTranslationMessages } from '../i18n';
+import { formatTranslationMessages } from '../../i18n';
 
-jest.mock('../translations/en.json', () => ({
+jest.mock('../../translations/en.json', () => ({
   message1: 'default message',
   message2: 'default message 2',
 }));
