+++ app/tests/boilerplate/utils/injectReducer.test.js
@@ -7,9 +7,9 @@ import { shallow } from 'enzyme';
 import React from 'react';
 import identity from 'lodash/identity';
 
-import configureStore from '../../configureStore';
-import injectReducer from '../injectReducer';
-import * as reducerInjectors from '../reducerInjectors';
+import configureStore from '../../../configureStore';
+import injectReducer from '../../../utils/injectReducer';
+import * as reducerInjectors from '../../../utils/reducerInjectors';
 
 // Fixtures
 const Component = () => null;
