+++ app/tests/boilerplate/utils/reducerInjectors.test.js
@@ -5,9 +5,11 @@
 import { memoryHistory } from 'react-router-dom';
 import identity from 'lodash/identity';
 
-import configureStore from '../../configureStore';
+import configureStore from '../../../configureStore';
 
-import getInjectors, { injectReducerFactory } from '../reducerInjectors';
+import getInjectors, {
+  injectReducerFactory,
+} from '../../../utils/reducerInjectors';
 
 // Fixtures
