+++ app/tests/boilerplate/utils/request.test.js
@@ -2,7 +2,7 @@
  * Test the request function
  */
 
-import request from '../request';
+import request from '../../../utils/request';
 
 describe('request', () => {
   // Before each test, stub the fetch function
